VTKU-51 : Extract parameter handling code
No functional changes, I hope :)
