+++ sijoittelu-service/src/test/java/fi/vm/sade/sijoittelu/SijoitteluIntegrationTestToBeRunManually.java
@@ -15,6 +15,7 @@ import fi.vm.sade.sijoittelu.domain.SijoitteluAjo;
 import fi.vm.sade.sijoittelu.laskenta.external.resource.VirkailijaValintaTulosServiceResource;
 import fi.vm.sade.sijoittelu.laskenta.resource.SijoitteluResource;
 import fi.vm.sade.sijoittelu.laskenta.service.business.SijoitteluBusinessService;
+import fi.vm.sade.sijoittelu.laskenta.service.business.SijoitteluajoResourcesLoader;
 import fi.vm.sade.sijoittelu.laskenta.service.business.ValintarekisteriService;
 import fi.vm.sade.sijoittelu.laskenta.service.it.TarjontaIntegrationService;
 import fi.vm.sade.sijoittelu.tulos.dto.HakukohdeDTO;
@@ -155,7 +156,11 @@ public class SijoitteluIntegrationTestToBeRunManually {
 
         @Autowired
         public LightWeightSijoitteluBusinessServiceForTesting(SijoitteluTulosConverter sijoitteluTulosConverter, TarjontaIntegrationService tarjontaIntegrationService, VirkailijaValintaTulosServiceResource valintaTulosServiceResource, ValintarekisteriService valintarekisteriService) {
-            super(sijoitteluTulosConverter, tarjontaIntegrationService, valintaTulosServiceResource, valintarekisteriService, new SijoitteluConfiguration());
+            super(sijoitteluTulosConverter,
+                valintaTulosServiceResource,
+                valintarekisteriService,
+                new SijoitteluConfiguration(),
+                new SijoitteluajoResourcesLoader(tarjontaIntegrationService));
         }
 
         /**
@@ -179,7 +184,7 @@ public class SijoitteluIntegrationTestToBeRunManually {
             LOG.info(nameOfThisFakeSijoitteluRun + " sijoittelun koko: 0 olemassaolevaa, {} uutta, 0 valintatulosta", uudetHakukohteet.size());
             stopWatch.start("Luodaan sijoitteluajoWrapper ja asetetaan parametrit");
             final SijoitteluajoWrapper sijoitteluajoWrapper = SijoitteluajoWrapperFactory.createSijoitteluAjoWrapper(new SijoitteluConfiguration(), sijoitteluAjo, uudetHakukohteet, Collections.emptyList(), Collections.emptyMap());
-            asetaSijoittelunParametrit(hakuOid, sijoitteluajoWrapper, findParametersFromTarjontaAndPerformInitialValidation(hakuOid));
+            sijoitteluajoResourcesLoader.asetaSijoittelunParametrit(hakuOid, sijoitteluajoWrapper, sijoitteluajoResourcesLoader.findParametersFromTarjontaAndPerformInitialValidation(hakuOid));
             stopWatch.stop();
 
             LOG.info("Suoritetaan " + nameOfThisFakeSijoitteluRun);
